move dependencies of test on state, evm and account to manager/erismint
